+++ src/main/java/com/gdmc/httpinterfacemod/commands/SetBuildAreaCommand.java
@@ -1,4 +1,4 @@
-package com.gdmc.httpinterfacemod.settlementcommand;
+package com.gdmc.httpinterfacemod.commands;
 
 import com.gdmc.httpinterfacemod.handlers.BuildAreaHandler;
 import com.mojang.brigadier.CommandDispatcher;
@@ -6,26 +6,28 @@ import com.mojang.brigadier.context.CommandContext;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.arguments.coordinates.BlockPosArgument;
+import net.minecraft.commands.arguments.coordinates.Coordinates;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
-public class SetBuildAreaCommand<S> {
+public final class SetBuildAreaCommand {
 
     private static final String COMMAND_NAME = "setbuildarea";
-    private static final Logger LOGGER = LogManager.getLogger();
 
     private SetBuildAreaCommand() { }
 
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
-        dispatcher.register(
-                Commands.literal(COMMAND_NAME)
-                .then(Commands.argument("from", BlockPosArgument.blockPos())
-                .then(Commands.argument("to", BlockPosArgument.blockPos())
-                .executes( context -> {
-                    return perform(context, BlockPosArgument.getLoadedBlockPos(context, "from"), BlockPosArgument.getLoadedBlockPos(context, "to"));
-                }))));
+        dispatcher.register(Commands.literal(COMMAND_NAME)
+            .then(
+                Commands.argument("from", BlockPosArgument.blockPos())
+            .then(
+                Commands.argument("to", BlockPosArgument.blockPos())
+            .executes( context -> perform(
+                context,
+                context.getArgument("from", Coordinates.class).getBlockPos(context.getSource()),
+                context.getArgument("to", Coordinates.class).getBlockPos(context.getSource())
+            ))))
+        );
     }
 
     private static int perform(CommandContext<CommandSourceStack> commandSourceContext, BlockPos from, BlockPos to) {
@@ -39,7 +41,6 @@ public class SetBuildAreaCommand<S> {
         BuildAreaHandler.setBuildArea(x1, y1, z1, x2, y2, z2);
         String feedback = String.format("Build area set to %d, %d, %d to %d, %d, %d,", x1, y1, z1, x2, y2, z2 );
         commandSourceContext.getSource().sendSuccess(Component.nullToEmpty(feedback), true);
-        LOGGER.info(feedback);
         return 1;
     }
 }