+++ src/main/java/com/gdmc/httpinterfacemod/handlers/CommandHandler.java
@@ -7,6 +7,7 @@ import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.world.phys.Vec3;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -27,13 +28,26 @@ public class CommandHandler extends HandlerBase {
 
         Map<String, String> queryParams = parseQueryString(httpExchange.getRequestURI().getRawQuery());
 
-        String dimension = queryParams.getOrDefault("dimension", null);
+        // POST: x, y, z positions
+        int x;
+        int y;
+        int z;
+        String dimension;
+        try {
+            x = Integer.parseInt(queryParams.getOrDefault("x", "0"));
+            y = Integer.parseInt(queryParams.getOrDefault("y", "0"));
+            z = Integer.parseInt(queryParams.getOrDefault("z", "0"));
+
+            dimension = queryParams.getOrDefault("dimension", null);
+        } catch (NumberFormatException e) {
+            throw new HttpException("Could not parse query parameter: " + e.getMessage(), 400);
+        }
 
         // execute command(s)
         InputStream bodyStream = httpExchange.getRequestBody();
         List<String> commands = new BufferedReader(new InputStreamReader(bodyStream)).lines().toList();
 
-        CommandSourceStack cmdSrc = createCommandSource("GDMC-CommandHandler", dimension);
+        CommandSourceStack cmdSrc = createCommandSource("GDMC-CommandHandler", dimension, new Vec3(x, y, z));
 
         JsonArray returnValues = new JsonArray();
         for (String command: commands) {