+++ CHANGELOG.md
@@ -1,4 +1,14 @@
+# GDMC-HTTP 1.1.0 (Minecraft 1.19.2)
+
+- NEW: `POST /command` now accepts `x`, `y`, `z` parameters, usefull for when using commands with [relative coordinates](https://minecraft.fandom.com/wiki/Coordinates#Relative_world_coordinates) or commands such as [/locate](https://minecraft.fandom.com/wiki/Coordinates#Relative_world_coordinates).
+- NEW: Add `OPTIONS /` to get version of Minecraft and version of GDMC-HTTP interface.
+- NEW: Port number of the HTTP interface can be changed using the `/sethttpport <port>` Minecraft console command. This value will be saved to a config file and therefore will be persistent.
+- NEW: Get port number of the HTTP interface via the `/gethttpport` Minecraft console command.
+- FIX: The `/setbuildarea` console command can now also accept coordinates that aren't loaded yet.
+- FIX: Explicitly set HTTP interface address to that of the "loop back" address, which is usually "localhost".
+
 # GDMC-HTTP 1.0.0 (Minecraft 1.19.2)
+
 - BREAKING: JSON-formatted NBT-like data is no longer supported in request bodies. Use [SNBT notation](https://minecraft.fandom.com/wiki/NBT_format#SNBT_format) instead.
 - BREAKING: Properties containing NBT values in JSON responses are no longer formatted as JSON, but as [SNBT strings](https://minecraft.fandom.com/wiki/NBT_format#SNBT_format).
 - BREAKING: Plain-text formatted responses have been removed in favour of JSON.
@@ -7,10 +17,12 @@
 - FIX: Improved performance!
 
 # GDMC-HTTP 0.7.6 (Minecraft 1.19.2)
+
 - FIX: `GET /biomes` now returns an empty string for the biome ID if the requested position is outside of the vertical boundaries of the world.
 - FIX: Typo in error message `POST /structure` handler.
 
 # GDMC-HTTP 0.7.5 (Minecraft 1.19.2)
+
 - NEW: `GET /entities` for reading entities within a certain area.
 - NEW: `PUT /entities` for creating any number of entities.
 - NEW: `PATCH /entities` for editing existing entities in the world.
@@ -20,18 +32,23 @@
 - FIX: Partial refactor for improved readability and decreased branching.
 
 # GDMC-HTTP 0.7.4 (Minecraft 1.19.2)
+
 - FIX: Placement of multi-part blocks such as beds and doors.
 
 # GDMC-HTTP 0.7.3 (Minecraft 1.19.2)
+
 - NEW: Improved user-facing documentation.
 
 # GDMC-HTTP 0.7.2 (Minecraft 1.19.2)
+
 - FIX: Add proper exception handling for malformed JSON input at `PUT /blocks`.
 
 # GDMC-HTTP 0.7.1 (Minecraft 1.19.2)
+
 - FIX: Allow endpoint `PUT /blocks` to process block placement instructions without (valid) coordinates. It places it at the URL query coordinates instead.
 
 # GDMC-HTTP 0.7.0 (Minecraft 1.19.2)
+
 - NEW:`PUT /blocks` endpoint can now accept a JSON-formatted request body as input using the request header `"Content-Type": "application/json"`. The format is identical the response of the `GET /blocks` with the `"Accept": "application/json"` request header, with a few minor additional features:
   - For each placement instruction, `x`, `y` and `z` are optional. If omitted, it uses the coordinate set in the request's URL query as a fallback, or 0 if these aren't set.
   - For each placement instruction, `x`, `y` and `z` can be a integer ("x": `32`) or use tilde (`"x": "~32"`) or caret (`"x": "^32"`) notation in a string.
@@ -48,23 +65,29 @@
 - FIX: Minor improvements to performance.
 
 # GDMC-HTTP 0.6.5 (Minecraft 1.19.2)
+
 - NEW: Add `GET /structure` endpoint, for generating an NBT-formatted file of an area of blocks in the world.
 - NEW: Add `POST /structure` endpoint, for placing an NBT-formatted file of a structure into the world.
 
 # GDMC-HTTP 0.6.4 (Minecraft 1.19.2)
+
 - NEW: Enable `GET /chunks` to return a JSON-formatted response when the request header `Accept` is set to `application/json`.
 
 # GDMC-HTTP 0.6.3 (Minecraft 1.19.2)
+
 - NEW: Allow negative values for range coordinates in all endpoints that take the `dx`, `dy` and `dz` parameters.
 
 # GDMC-HTTP 0.6.2 (Minecraft 1.19.2)
+
 - FIX: Fixed formatting coordinates in plain-text response for `GET /blocks` endpoint.
 
 # GDMC-HTTP 0.6.1 (Minecraft 1.19.2)
+
 - NEW: Add `GET /biomes` endpoint to request biome of a location in the world.
 - FIX: Minor performance improvements.
 
 # GDMC-HTTP 0.6.0 (Minecraft 1.19.2)
+
 - NEW: `PUT /blocks` can now also process block entity data (eg. chest contents, a book on a lectern, armor on an armor stand, etc.).
 - NEW: `GET /blocks` can return multiple blocks within a given area.
 - NEW: Response of `GET /blocks` now always includes a block's position in front of its material.
@@ -75,17 +98,21 @@
 - FIX: Minor performance improvements.
 
 # GDMC-HTTP 0.5.3 (Minecraft 1.19.2)
+
 - NEW: Implement parameter for specify target dimension (overworld, nether, end, etc.) for the `/blocks`, `/command` and `/chunks` endpoints. For example `PUT /blocks?dimension=nether` places blocks in the nether instead of the overworld. If parameter is omitted it defaults to the overworld.
 - NEW: Implement `OPTIONS` method for the `/blocks` endpoint.
 - FIX: Relative position placement for blocks with a `~<int>` position when placing blocks with the `PUT /blocks` endpoint.
 
 # GDMC-HTTP 0.5.2 (Minecraft 1.19.2)
+
 - BREAKING: Renamed `GET /info` endpoint to `GET /version`.
 
 # GDMC-HTTP 0.5.1 (Minecraft 1.19.2)
+
 - NEW: Add `GET /info` endpoint, to get the current Minecraft version.
 
 # GDMC-HTTP 0.5.0 (Minecraft 1.19.2)
+
 - NEW: Compatibility with Minecraft version 1.19.2.
 - BREAKING: No longer compatible with versions of Minecraft other than 1.19.2.